agentic_framework/nova_agents/requirements.txt
google-cloud-aiplatform>=1.38.0
google-adk>=0.1.0
google-genai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio